+++ Model/Request/UpdateOrder.php
@@ -150,7 +150,7 @@ class UpdateOrder extends AbstractRequest implements RequestInterface
                 'merchantDetails'   => [
                     'customField1'  => $amount,
                     'customField2'  => json_encode($subs_data),
-                    'customField3'  => $this->config->getReservedOrderId($quoteId),
+                    //'customField3'  => $this->config->getReservedOrderId($quoteId),
                     // customField4 will be set in AbstractRequest class
                     'customField5' => $currency,
                 ],