+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Magento 2 Nuvei Checkout Module
 ---
 
+# 3.1.4-p2
+```
+    * For the magento v2.4.* versions where Zend\Uri\Uri class is missing. It is replaced with Laminas\Uri\Uri.
+    * Fixed the bug when try to set custom title on Nuvei Subscriptions page.
+```
+
 # 3.1.4-p1
 ```
     * Added new SDK's URLs for the QA site.